[refac] sidebar 컴포넌트 커스텀 훅 사용해서 리팩토링
